+++ infrastructure/datastore/migrations/20211216_initialcreate.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE beer (
     name varchar(100) NOT NULL,
     brewery varchar(100) NOT NULL,
     country varchar(100) NOT NULL,
-    price  bigint,
+    price  float,
     currency varchar(100) NOT NULL 
 );
